+++ src/main/java/inshining/virtualstorage/service/storage/FolderStorageService.java
@@ -1,4 +1,4 @@
-package inshining.virtualstorage.service;
+package inshining.virtualstorage.service.storage;
 
 public interface FolderStorageService {
     boolean createFolder(String username, String storagePath);
