+++ src/main/java/inshining/virtualstorage/service/storage/StorageService.java
@@ -1,8 +1,7 @@
-package inshining.virtualstorage.service;
+package inshining.virtualstorage.service.storage;
 
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
