+++ src/test/java/inshining/virtualstorage/metadata/service/FileMetaDataServiceTest.java
@@ -4,8 +4,8 @@ import inshining.virtualstorage.dto.FileDownloadDTO;
 import inshining.virtualstorage.dto.SuccessResponse;
 import inshining.virtualstorage.model.FileMetaData;
 import inshining.virtualstorage.service.FileService;
-import inshining.virtualstorage.service.LocalStorageService;
-import inshining.virtualstorage.service.FileMetaDataService;
+import inshining.virtualstorage.service.storage.LocalStorageService;
+import inshining.virtualstorage.service.metadata.FileMetaDataService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
