+++ src/test/java/inshining/virtualstorage/metadata/service/FolderMetaDataServiceTest.java
@@ -5,8 +5,8 @@ import inshining.virtualstorage.dto.FolderCreateResponse;
 import inshining.virtualstorage.model.FileMetaData;
 import inshining.virtualstorage.model.FolderMetaData;
 import inshining.virtualstorage.repository.FakeFolderMetaDataRepository;
-import inshining.virtualstorage.service.FileMetaDataService;
-import inshining.virtualstorage.service.FolderMetaDataService;
+import inshining.virtualstorage.service.metadata.FileMetaDataService;
+import inshining.virtualstorage.service.metadata.FolderMetaDataService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
