+++ src/test/java/inshining/virtualstorage/metadata/service/FolderServiceTest.java
@@ -6,9 +6,9 @@ import inshining.virtualstorage.dto.FolderMetaResponse;
 import inshining.virtualstorage.model.FileMetaData;
 import inshining.virtualstorage.model.FolderMetaData;
 import inshining.virtualstorage.repository.FakeFolderMetaDataRepository;
-import inshining.virtualstorage.service.FileMetaDataService;
-import inshining.virtualstorage.service.FolderLocalStorageService;
-import inshining.virtualstorage.service.FolderMetaDataService;
+import inshining.virtualstorage.service.metadata.FileMetaDataService;
+import inshining.virtualstorage.service.storage.FolderLocalStorageService;
+import inshining.virtualstorage.service.metadata.FolderMetaDataService;
 import inshining.virtualstorage.service.FolderService;
 import inshining.virtualstorage.util.FileDeletor;
 import org.junit.jupiter.api.Assertions;
